Multi-threaded test and moved all test into one file
